Finish Init Function

+++ SPI_Driver/Common_Macros.h
@@ -17,7 +17,7 @@
 #define SET_BIT(REG,BIT) (REG|=(1<<BIT))
 
 /* Clear a certain bit in any register */
-#define CLEAR_BIT(REG,BIT) (REG&=(~(1<<BIT)))
+#define CLEAR_BIT(REG,BIT) (REG&=(uint32)(~(1<<BIT)))
 
 /* Toggle a certain bit in any register */
 #define TOGGLE_BIT(REG,BIT) (REG^=(1<<BIT))
